requirements.txt
astroid==2.3.3
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-ckeditor==5.8.0
django-countries==5.5
django-crispy-forms==1.8.0
django-filter==2.2.0
django-js-asset==1.2.2
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.2.1
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
stripe==2.41.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
